+++ NEWS.md
@@ -1,5 +1,8 @@
 # DIscBIO (development version)
 
+* Delinted package
+* Removed dependency on `philentropy` and `org.Hs.eg.db` (issue #44, #45)
+
 # DIscBIO 1.2.2
 
 * Fixed calculation of Jaccard index (issue #44)